% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipeds_coordinates}
\alias{ipeds_coordinates}
\title{IPEDS Coordinates}
\format{
Data frame with columns
\describe{
\item{unit_id}{IPEDS unit ID number.}
\item{institution_name}{Name of school.}
\item{sector_description}{Sector of institution.}
\item{state_description}{State of institution.}
\item{latitude, longitude}{Latitude and longitude of institution.}
}
}
\source{
IPEDS Data Center,
 <https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx?goToReportId=7>
}
\usage{
ipeds_coordinates
}
\description{
Latitude and longitude for IPEDS institutions in the Lower 48 states.
}
\keyword{datasets}
