% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipeds_directory}
\alias{ipeds_directory}
\title{IPEDS Directory}
\format{
Data frame with columns
\describe{
\item{unit_id}{IPEDS unit ID number.}
\item{institution_name}{Name of school.}
\item{year}{Year of data.}
\item{sector_description}{Sector of institution}
\item{state_description}{State of institution.}
\item{latitude, longitude}{Latitude and longitude of institution.}
}
}
\source{
IPEDS Data Center,
 <https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx?goToReportId=7>
}
\usage{
ipeds_directory
}
\description{
IPEDS directory information.
}
\keyword{datasets}
