% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipeds_distance}
\alias{ipeds_distance}
\title{IPEDS Distance}
\format{
Data frame with columns
\describe{
\item{unit_id}{IPEDS unit ID number.}
\item{institution_name}{Name of school.}
\item{year}{Year of data.}
\item{sector_description}{Sector of institution.}
\item{level}{Student level.}
\item{distance_type}{No distance, some distance, or all distance.}
\item{headcount}{Number of students.}
}
}
\source{
IPEDS Data Center,
 <https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx?goToReportId=7>
}
\usage{
ipeds_distance
}
\description{
IPEDS distance education headcounts.
}
\keyword{datasets}
