% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipeds_funnel}
\alias{ipeds_funnel}
\title{IPEDS Funnel}
\format{
Data frame with columns
\describe{
\item{unit_id}{IPEDS unit ID number.}
\item{institution_name}{Name of school.}
\item{sector_description}{Sector of institution.}
\item{state_description}{State of institution.}
\item{ivy_plus_exchange}{Indicates Ivy League and IvyPlus Exchange schools.}
\item{applicants_total}{Number of first-time in college applicants.}
\item{admissions_total}{Number of first-time in college applicants who were admitted.}
\item{enrolled_total}{Number of first-time in college applicants who enrolled.}
\item{admit_rate}{The proportion of applicants who were admitted.}
\item{yield_rate}{The proportion of admits who enrolled.}
}
}
\source{
IPEDS Data Center,
 <https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx?goToReportId=7>
}
\usage{
ipeds_funnel
}
\description{
IPEDS first-time in college applied, admitted, enrolled data for 4-year
public and 4-year private not-for-profit schools.
}
\keyword{datasets}
