% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipeds_salary}
\alias{ipeds_salary}
\title{IPEDS Salary}
\format{
Data frame with columns
\describe{
\item{unit_id}{IPEDS unit ID number.}
\item{institution_name}{Name of school.}
\item{year}{Year of data.}
\item{sector_description}{Sector of institution.}
\item{state_description}{State of institution.}
\item{headcount}{Number of employees.}
\item{salary}{Total salary.}
}
}
\source{
IPEDS Data Center,
 <https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx?goToReportId=7>
}
\usage{
ipeds_salary
}
\description{
IPEDS number and salary outlays for full-time nonmedical noninstructional staff by occupation.
}
\keyword{datasets}
