% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipeds_sat}
\alias{ipeds_sat}
\title{IPEDS Test Scores}
\format{
Data frame with columns
\describe{
\item{unit_id}{IPEDS unit ID number.}
\item{institution_name}{Name of school.}
\item{percentile}{SAT score percentile.}
\item{sat_ebrw}{SAT Evidenced Based Reading score.}
\item{sat_math}{SAT Math score.}
}
}
\source{
IPEDS Data Center,
 <https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx?goToReportId=7>
}
\usage{
ipeds_sat
}
\description{
IPEDS SAT scores for first-time in college students
}
\keyword{datasets}
