% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipeds_sector}
\alias{ipeds_sector}
\title{IPEDS Sector}
\format{
Data frame with columns
\describe{
\item{sector_code, sector_description}{Sector code and sector description.}
}
}
\source{
IPEDS Data Center,
 <https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx?goToReportId=7>
}
\usage{
ipeds_sector
}
\description{
IPEDS sector code and sector description lookup table.
}
\keyword{datasets}
