% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipeds_state}
\alias{ipeds_state}
\title{IPEDS State}
\format{
Data frame with columns
\describe{
\item{state_abbreviation, state_description}{State code and state description.}
}
}
\source{
IPEDS Data Center,
 <https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx?goToReportId=7>
}
\usage{
ipeds_state
}
\description{
US state abbreviation and state description lookup table.
}
\keyword{datasets}
